sg: device timestamp
Add timestamp to each sg_device object that is written when the
object is created. The timestamp is the number nanoseconds since
the boot time of the machine.

The purpose is to allow the user to detect, via the extended
ioctl()s SG_SEIRV_DEV_TS_LOWER and SG_SEIRV_DEV_TS_UPPER, if
a given sg device object (e.g. /dev/sg3) may have possibly
changed. One worrisome scenario is when a device disappears
and a newly connected device takes the same sg device object
number (e.g. /dev/sg3) as the recently disappeared device.
Linux gives no guarantees that this type of behaviour will
_not_ happen. Recording the device creation timestamp is one
way an application can detect when this happens.

The uptime command in Linux shows, in humanly readable form,
how long a machine has been "up", that is the time that has
elapsed since the machine was started or "rebooted".

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>