sg: condition met is not an error
Most, but not all, non-zero SCSI status values indicate there is
a problem with the current or earlier command. The main exception
is from the PRE-FETCH commands that return CONDITION MET (0x4)
when the specified blocks can/have fitted in the device's cache.
And somewhat strangely return the GOOD (0x0) status if it will/has
not fit. Clean up all SCSI status and associated error
processing.

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>