sg: tweak sg_find_sfp_by_fd()
The sg_find_sfp_by_fd() function is called every time a file share
is established. If request sharing is being used to copy to two
or more destinations, there will be many calls to this function
to swap between those destination, so its performance may become
important. Simplify the "search" by drilling into the given
fd's 'struct file' as, if all is well, the wanted sfp is in
filp->private_data.

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>