sg: add no_attach_msg parameter
When testing and with big storage arrays having a log messaage
for each sg device attached becomes a little annoying. Still keep
that as the default but when no_attach_msg=1 is given at driver
or module load time, then any sg device nodes attached will not
be reported to the log.

Re-order the other three driver/module load time parameters so
they appear in alphabetical order when viewed with modinfo.

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>