sg: add SGV4_FLAG_REC_ORDER
By default, when ioctl(SG_IORECEIVE) is used in multiple requests
mode (mrq) the response array is built in completion order. And the
completion order isn't necessarily submission order which can be a
nuisance. This new flag allows the user to specify where (via an
index in the v4::request_priority field) a given request's response
will be placed in the response array associated with a mrq
ioctl(SG_IORECEIVE) call.

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>