scsi: target: core: Set OVERFLOW/UNDERFLOW bit for residual write command
According to RFC 7143 11.4.5.2.:

If SPDTL > EDTL for a task, iSCSI Overflow MUST be signaled in the
SCSI Response PDU as specified in Section 11.4.5.1.  The Residual
Count MUST be set to the numerical value of (SPDTL - EDTL).

If SPDTL < EDTL for a task, iSCSI Underflow MUST be signaled in the
SCSI Response PDU as specified in Section 11.4.5.1.  The Residual
Count MUST be set to the numerical value of (EDTL - SPDTL).

The libiscsi residual write tests checks whether the returned residual
amount is the same as it was intended by test. The target did not set
the residual amount and OVERFLOW/UNDERFLOW bit in the SCSI Response PDU
for DMA_TO_DEVICE commands that contains LBAs so the initiator did not
get the correct data and the libiscsi tests didn't pass. Such target
behavior didn't meet the RFC 7143 document.

Set the OVERFLOW/UNDERFLOW bit and the residual amount before
checking the data direction and the cmd flags so that the target returns
the correct value to the initiator.

Reviewed-by: Roman Bolshakov <r.bolshakov@yadro.com>
Signed-off-by: Anastasia Kovaleva <a.kovaleva@yadro.com>