scsi: target: core: Change ASCQ for residual write command
According to FCS-4 (9.4.2):
If the command requested that data beyond the length specified by the FCP_DL
field be transferred, then the device server shall set the FCP_RESID_OVER
bit (see 9.5.8) to one in the FCP_RSP IU and:

a) process the command normally except that data beyond the FCP_DL count
shall not be requested or transferred;

b) transfer no data and return CHECK CONDITION status with the sense key set
to ILLEGAL REQUEST and the additional sense code set to INVALID FIELD IN
COMMAND INFORMATION UNIT; or

c) may transfer data and return CHECK CONDITION status with the sense key set
to ABORTED COMMAND and the additional sense code set to INVALID FIELD IN
COMMAND INFORMATION UNIT.

The target returned INVALID_CDB_FIELD ASCQ, which did not meet the standard.
Some SCSI testing libraries like libiscsi checks the returned ASCQ accordig to
the standart. This checks did not pass.

Change the ASCQ to INVALID FIELD IN COMMAND INFORMATION UNIT to meet the
standart.

Reviewed-by: Roman Bolshakov <r.bolshakov@yadro.com>
Signed-off-by: Anastasia Kovaleva <a.kovaleva@yadro.com>