sg: blk_poll/hipri work for mrq
There are several variants here to cope with. The SGV4_FLAG_HIPRI
flag may be set on the control object and/or set on any number of
requests in the req_arr. To simplify things, if that flag is
set on the control object, then its is set internally on each
request in req_arr. That leaves the case were this flag is cleared
on the control object, but set on some or all of the requests in
req_arr. That will work but it may not be the best approach, see
the next paragraph.

For the shared variable blocking mrq method, if the SGV4_FLAG_HIPRI
flag is set on the control object, then an interleaved blk_poll()
algorithm is performed on inflight requests. This should give
better results than doing blk_poll() on each request until it is
completed before moving onto the next inflight request.

Combine SG_PACK_ID_WILDCARD and SG_TAG_WILDCARD into
SG_PACK_ID_TAG_WILDCARD since they convey the same meaning and have
the same value.

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>