sg: bump version to 4.0.47
Version is bumped to > 4.0.30 so test tools is sg3_utils (beta 1.45
revision 827 and later) can distinguish between this full
featured driver and a reduced featured version. The full featured
driver has request sharing and multiple requests (in a single
invocation) support. The reduced featured driver only adds sg v4
interface support and its version numbers
are >= 4.0.00 and < 4.0.30 .

Both versions of the sg v4 driver are described on this webpage:
    https://sg.danny.cz/sg/sg_v40.html

Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>