Add passthrough abort handling in qla2xxx and lpfc
When command is aborted, in order to have the ability to requeue a
commandi, qla2xxx set hostbyte DID_RESET, which is not correct
semantically. lpfc sets DID_REQUEUE hostbyte to immediately requeue
command, which makes it impossible to handle an ABORT TASK function
manually.  The hostbyte DID_ABORT sends a command into an error handler.
Introduce a new hostbyte to have an oppotunity to handle responce to an
aborted command.